use uint16_t

+++ src/Display.cpp
@@ -80,7 +80,7 @@ void Display::Toggle::readCheckState(DisplayCoordinates touchCoordinates, const 
 
 std::pair<DisplayCoordinates*, int> Display::readTouchData() {
     static DisplayCoordinates touchEvent;
-    int n = 0;
+    uint16_t n = 0;
     uint16_t x[2], y[2];
     if (display->tsAvailable()){
         n = display->tsGetData(x, y);
